Documentation formatting and cleanup
Previously, the documentation was automatically generated from the
Python source using pydocmd.  The resulting documentation was adequate,
but wasn't formatted in the most reasonable way.  Particluarly,
the named tuple documentation was presented oddly.

This is a big hand-edited formatting change of the documentation, and
also a lot of small scale text changes to the contents, for increased
clarity.