Check if command future is cancelled before setting exception
When the command task exits, it completes all command futures
with an exception.  This is done because without a command task,
the futures would otherwise never complete, and leave the tasks
awaiting the result of those commands hanging.

However, if we are cleaning up after several tasks at once, the
waiting command task may have already been cancelled while it
was awaiting the future, which means the future has already
been cancelled.  So, we need to first check whether the future
is already complete before setting the exception, to avoid
future.set_exception() from itself raising an exception
during task cancellation.