Exit the concurrent trace example when SPACEBAR is pressed
The concurrent trace will now exit only when the spacebar is pressed.
This has a few advantages.  First, the display of the trace persists
after it has completed, until the user is done with it.  Second,
it provides a better example of multiple concurrent operations, since
there is one task for handling user input, and another for performing
the trace.  It also exercises the cancellation path of mtrpacket.