Use gather to wait for spacebar and probe task
Rather than waiting only on the spacebar task, and then cancelling
the probe task, wait on both with gather, and have the spacebar
task cancel the probe task after spacebar is pressed.

This allows unexpected exceptions in the probe task to be immediately
reported, rather than waiting for spacebar to be pressed before
any exception is reported.  The previous behavior seemed strange,
with the script apparently stopping, but without an error.