Handle unresolving hostnames better in trace-concurrent.py
If a hostname fails to resolve, catch the HostResolveError and
print a resonable error message.  Cancel the probe task and spacebar
task in the case of an exception raised during execution, to avoid
warnings about pending tasks when the event loop is closed.

Also, display the hostname being traced to during normal execution.