Cache resolved IP addresses
Previously, when sending multiple probes to a particular hostname, the
hostname was resolved to an IP address each time a probe was sent.
When many concurrent probes are continuously sent (with hundreds
to thousands of probes) the DNS resolution had enough overhead to
have a measurable performance impact.

To mitigate this, the IP address are now cached after the first time
a hostname is resolved.

MtrPacket.clear_dns_cache() was added to clear that DNS cache for
an MtrPacket session.