Take an optional command argument to the MtrPacket constructor.
An argument can be specified to the MtrPacket constructor which will
specify a full shell command to use when invoking the 'mtr-packet'
subprocess.  Only if no argument is supplied will the MTR_PACKET
environment variable be used, and only if neither are present will
the default of 'mtr-packet' be used.

This addresses the request in PR #5.